Rework workflow
This makes immucore run more in parallel rather than block everything
else. We just tell sysroot.mount that eventually it will be mounted and
to wait for a bit.

This allows us to be more flexible where to run and run in parallel in
cases like cdrom in which we may do things but we need the sysroot to be
mounted already but not from us.

 - Dont create sysroot, just add a timeout override so it waits for us
 - Dont block on the service, just make sure to finish before
   initrd.target
 - Fix mounts from cmdline

Signed-off-by: Itxaka <itxaka.garcia@spectrocloud.com>