:bug: Retry getting the state label
On some devices, the disk can be a bit late to appear, so when getting
the state label and fs type it can lead to a direct failure.

This patch introduces a retry with a 1 second delay between tries in
order to obtain the state label. If that fails after all the retries it
also panics so we stop running immucore as it makes no sense.

Signed-off-by: Itxaka <itxaka.garcia@spectrocloud.com>