Upgrade kcrypt partitions on boot
This solves the issue of upgrading from the old method of mappings to
identify the kcrypt partitions, with the new one that uses predictable
uuids to identify them.

On boot we check for luks partitions that match the partitions persistent
label (NOT THE FS LABEL) as that gets stored properly.
If it matches persistent, we get its uuid and match it agsint the know
uuid. If they dont match, we update it to match and proceed with
booting.

Signed-off-by: Itxaka <itxaka.garcia@spectrocloud.com>