Try to not mount COS_OEM before decryption happens (#509)
* [WIP] Try to not mount COS_OEM before decryption happens

because COS_OEM could be encrypted.

Part of: https://github.com/kairos-io/kairos/issues/2988

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>

* Make oem mount depend on kcrypt unlock

because the OEM partition may be locked

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>

* Unlock partitions using the new Encryptor interface

the same way the kairos-agent encrypts the partitions

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>

* Use latest (local) checkouts of deps to fix conflic (will bump before merge)

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>

* Setup networking before reaching out to the KMS

We need a proper fix for the dns. Setting up google dns servers is not a
proper fix.

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>

* Remove replace directives from go.mod

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>

* Disable UKI network setup for KMS

Until we can get the network setup right. For now only PCR locking for
UKI.

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>

* Bump kairos-sdk

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>

* Skip non-used method early (just in case someone tries to call it)

We need to fix the network setup before we enable this. This is the only
thing left to get KMS working in UKI mode.

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>

* Mount OEM before kcrypt if OEM is not locked

because in old installations, grub won't have the extended cmdline (in
the state partition) which makes sure kcrypt configuration is available
before we mount OEM. To avoid breaking things, if OEM is not encrypted,
we will just mount OEM like we used to (before kcrypt)

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>

* Bump to tagged version of kairos-sdk

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>

---------

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>